+++ psyrun/backend/tests/test_load_balancing.py
@@ -7,8 +7,13 @@ from psyrun.store.npz import NpzStore
 from psyrun.store.pickle import PickleStore
 
 
-def square(a):
-    return {'x': a ** 2}
+def square(a, p):
+    return {'x': a ** p}
+
+
+def setup_fn(i):
+    assert i == 0
+    return {'p': 2}
 
 
 @pytest.mark.parametrize('store', [PickleStore(), NpzStore(), H5Store()])
@@ -18,8 +23,8 @@ def test_load_balancing_worker(tmpdir, store):
     statusfile = str(tmpdir.join('status'))
     store.save(infile, Param(a=range(7)).build())
     LoadBalancingWorker.create_statusfile(statusfile)
-    worker = LoadBalancingWorker(infile, outfile, statusfile, store)
-    worker.start(square)
+    worker = LoadBalancingWorker(0, infile, outfile, statusfile, store)
+    worker.start(square, setup_fn)
     result = store.load(outfile)
     assert sorted(result['a']) == sorted(range(7))
     assert sorted(result['x']) == [i ** 2 for i in range(7)]